fix: outfile
