+++ srcs/minishell_execution.c
@@ -6,108 +6,11 @@
 /*   By: sharnvon <sharnvon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/29 20:38:04 by sharnvon          #+#    #+#             */
-/*   Updated: 2022/10/11 21:21:37 by sharnvon         ###   ########.fr       */
+/*   Updated: 2022/10/12 01:06:21 by sharnvon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-pid_t	ppid;
-
-int	execution_path_command(t_shell *shell, char **command)
-{
-	int		index;
-	char	*path;
-	char	**env_path;
-
-// * check not real path should path in shell->env
-	env_path = ft_split_mode(getenv("PATH"), ':', BOUND);  //? FREE ?? //
-
-	index = 0;
-	while (env_path[index] != NULL)
-	{
-		path = ft_midjoin(env_path[index++], command[0], '/');
-		//* command is exist and can execute */
-		if (access(path, F_OK | R_OK | X_OK) == 0)
-		{
-
-			execution_token(shell, path, command);
-			if (shell->exstat == -1)
-			{
-				free_double_pointer(NULL, env_path, path);
-				return (-1);
-			}
-			break ;
-		}
-		//* command is not exist and can't execute */
-		else if (env_path[index + 1] == NULL)
-		{
-			printf("minishell: command not found: %s\n", command[0]);
-			shell->exstat = 127;		
-		}
-	}
-	free_double_pointer(NULL, env_path, path);
-	return (0);
-}
-
-int	execution_token(t_shell *shell, char *path, char **command)
-{
-	pid_t	pid;
-
-	pid = fork();
-	if (pid < 0)
-		return (-1);
-	else if (pid == 0)
-	{
-		execve(path, command, NULL);
-	}
-	else if (pid > 0)
-	{
-		waitpid(pid, &shell->exstat, 0);
-	}
-	return (0);
-}
-
-int	redirect_dup_input(int *fd)
-{
-	int		infile;
-
-	infile = open(IN_FILE, O_RDONLY);
-	if (infile < 0)
-	{
-		close(fd[0]);
-		return (0);
-	}
-	dup2(infile, 0);
-	close(infile);
-	close(fd[0]);
-	return (0);
-}
-
-int	redirect_dup_start(t_shell *shell, t_execute *exe)
-{
-	int		file;
-	t_cmd	*cmd;
-
-	file = 0;
-	cmd = (t_cmd *)shell->cmds->content;
-	redirect_dup_input(exe->fd);
-	if (cmd->fg == NULL && cmd->fgg == NULL && cmd->opt != 0)
-		dup2(exe->fd[1], 1);
-	else if (cmd->fg != NULL && cmd->fg[exe->index] != NULL)
-		file = open(cmd->fg[exe->index++], O_RDWR | O_CREAT, 0644);
-	else if ((cmd->fgg != NULL && cmd->fgg[exe->xedni] != NULL))
-		file = open(cmd->fgg[exe->xedni++], O_RDWR | O_CREAT | O_APPEND, 0644);
-	if (file < -1)
-		return (-1);
-	else if (file > 2)
-	{
-		dup2(file, 1);
-		close(file);
-	}
-	close(exe->fd[1]);
-	return (0);
-}
 
 void	execution_signal_handler(int signum)
 {
@@ -152,79 +55,23 @@ void	execution_signal(t_shell *shell, int mode)
 	}
 }
 
-int	execution_change_directory(t_shell *shell, char **command)
+void	execution_increasement(t_shell *shell,t_execute *exe)
 {
-	char	*user;
+	t_cmd *cmd;
 
-	user = ft_midjoin("~", getenv("USER"), '\0');
-	if (ft_lencount(NULL, command, STRS) == 3)
+	cmd = ((t_cmd *)shell->cmds->content);
+	if (cmd->fg != NULL && cmd->fg[exe->index] != NULL)
 	{
-		printf("cd: string not in pwd: %s\n", command[1]);
-		shell->exstat = 1;
+		exe->index++;
 	}
-	else if (ft_lencount(NULL, command, STRS) == 1 || string_compare(command[1], "~", NO_LEN) == 1 || string_compare(command[1], user, NO_LEN) == 1)
-		shell->exstat = chdir(getenv("HOME"));
-	else if (string_compare(command[1], "-", NO_LEN) == 1)
-		shell->exstat = chdir(getenv("OLDPWD"));
-	else if (ft_lencount(NULL, command, STRS) == 2)
-		shell->exstat = chdir(command[1]);
+	else if ((cmd->fgg != NULL && cmd->fgg[exe->xedni] != NULL))
+	{
+		exe->xedni++;
+	}
 	else
 	{
-		printf("cd: too many agurments");
-		shell->exstat = 1;
+		exe->index++;
 	}
-	free(user);
-	return (0);
-}
-
-// char	*ft_stringvalue(char *str)
-// {
-// 	int		index;
-// 	char	*result;
-
-// 	index = 0;
-// 	result = (char *)ft_calloc(sizeof(char), ft_lencount(str, NULL, STR) + 1);
-// 	if (result == NULL)
-// 		return (0);
-// 	while (str[index] != '\0')
-// 	{
-// 		result[index] = str[index];
-// 		index++;
-// 	}
-// 	result[index] = '\0';
-// 	return (result);
-// }
-
-int	execution_command(t_shell *shell, t_execute *exe)
-{	
-	t_cmd	*cmds;
-
-	cmds = exe->cmds;
-	execution_signal(shell, CHILD);
-	if (exe->index == 0 && exe->xedni == 0)
-	{
-		if (redirect_infile(shell, exe->cmds) < 0)
-			return (-1);
-	}
-	tcsetattr(STDIN_FILENO, TCSAFLUSH, &shell->terminal->shell);
-	sigaction(SIGQUIT, &shell->sigquit, NULL);
-	redirect_dup_start(shell, exe);
-	// if (string_compare(cmds->tokens[0], "$?", NO_LEN) == 1) //! check agign //
-	// 	shell->exstat = 127;
-	if (string_compare(cmds->tokens[0], "cd", NO_LEN) == 1)
-		shell->exstat = 0;
-	else if (string_compare(cmds->tokens[0], "export", NO_LEN) == 1)
-		shell->exstat = 0;
-	else if (string_compare(cmds->tokens[0], "unset", NO_LEN) == 1)
-		shell->exstat = 0;
-	else if (string_compare(cmds->tokens[0], "env", NO_LEN) == 1)
-		shell->exstat = execution_print_env(shell);
-	else if (access(cmds->tokens[0], F_OK | R_OK | X_OK) == 0)
-		shell->exstat = execution_token(shell, cmds->tokens[0], cmds->tokens);
-	else
-		execution_path_command(shell, cmds->tokens);
-	tcsetattr(STDIN_FILENO, TCSAFLUSH, &shell->terminal->minishell);
-	exit(shell->exstat);
 }
 
 int	execution_waitpid(t_shell *shell, t_execute *exe)
@@ -250,7 +97,8 @@ int	execution_waitpid(t_shell *shell, t_execute *exe)
 		dup2(shell->sinput, 0);
 	close(exe->fd[0]);
 	close(exe->fd[1]);
-	if (exe->index == 0 && exe->xedni == 0)
+	execution_increasement(shell, exe);
+	if (exe->index + exe->xedni <= exe->files)
 		return (1);
 	return (0);
 }
@@ -263,6 +111,7 @@ void	executeion_inite(t_shell *shell, t_execute *exe)
 	exe->cmds = (t_cmd *)shell->cmds->content;
 	exe->files = ft_lencount(NULL, exe->cmds->fg, STRS);
 	exe->files += ft_lencount(NULL, exe->cmds->fgg, STRS);
+	printf("files %d\n", exe->files);
 }
 
 int	cmd_execution(t_shell *shell)
